Fix an issue that would cause newly created items to have empty formula groups
When the item is first created, it would create an empty formula group, and the item would not have any formulae
Once the first formula was added, it would not automatically be added to a formula group.
This change has the following effect: if an item has formula groups but none of them contain any formulae, add all damage formulae to the first group.
It also has the effect of preventing a user from unassigning all formulae from all formula groups.